Delete datetime; js_only.
