+++ assets/prototype/domain/eventEditor/containers/mutations/dates.js
@@ -1,0 +1,15 @@
+import gql from 'graphql-tag';
+
+export const CREATE_DATE = gql`
+	mutation createDatetime($input: CreateDatetimeInput!) {
+		createDatetime(input: $input) {
+			datetime {
+				id
+				name
+				description
+				startDate
+				endDate
+			}
+		}
+	}
+`;
