+++ assets/prototype/domain/eventEditor/containers/mutations/useDeleteDateMutation.js
@@ -1,7 +1,8 @@
+import gql from 'graphql-tag';
 import { useMutation } from '@apollo/react-hooks';
 
 export const DELETE_DATETIME = gql`
-	mutation deleteDatetime($input: DeleteDatetimeInput!) {
+	mutation deleteDatetime($input: DeleteDatetimeInput) {
 		deleteDatetime(input: $input) {
 			datetime {
 				id
@@ -10,8 +11,10 @@ export const DELETE_DATETIME = gql`
 	}
 `;
 
-const useDeleteDatetimeMutation = ({ id }) => {
-	const [deleteDatetime] = useMutation(DELETE_DATETIME);
+const useDeleteDatetimeMutation = () => {
+	const [deleteDatetime, { loading, error }] = useMutation(DELETE_DATETIME);
+
+	console.log({ loading, error });
 
 	return deleteDatetime;
 };
