+++ assets/prototype/domain/eventEditor/containers/queries/dates.js
@@ -1,19 +1,5 @@
 import gql from 'graphql-tag';
 
-export const CREATE_DATE = gql`
-	mutation createDatetime($input: CreateDatetimeInput!) {
-		createDatetime(input: $input) {
-			datetime {
-				id
-				name
-				description
-				startDate
-				endDate
-			}
-		}
-	}
-`;
-
 export const GET_DATETIME = gql`
 	query DateTime($id: ID!) {
 		datetime(id: $id) {
